update to Simulation Map+Customer. Seems slower. Why?

+++ src/optimizer/CustomerPropositionDetails.cs
@@ -11,7 +11,7 @@ namespace optimizer
     {
         public string CustomerName { get; set; }
         public double ScoreContribution { get; set; }
-        public decimal LoanAmount { get; set; }
+        public double LoanAmount { get; set; }
         public double OptimalInterestRate { get; set; }
         public int OptimalMonthsPayBack { get; set; }
